+++ backend/packages/app/src/windup_app/server/media/__init__.py
@@ -1,5 +1,7 @@
 """媒体文件领域。"""
 
-from windup_app.server.media.model import MediaCategory, MediaUploadInput, MediaUploadResult
+from windup_common.enums.media import MediaCategory
+
+from windup_app.server.media.model import MediaUploadInput, MediaUploadResult
 
 __all__ = ["MediaCategory", "MediaUploadInput", "MediaUploadResult"]
