--- backend/packages/framework/src/windup_framework/storage/__init__.py
@@ -1,5 +1,0 @@
-"""对象存储技术能力。"""
-
-from windup_framework.storage.kodo import KodoStorage
-
-__all__ = ["KodoStorage"]
